requirements.txt
wxPython
pywebview
pyyaml
pyinstaller
